Ajustes e Novos Exercicios

+++ ex008.py
@@ -1,6 +1,6 @@
 # Escreva um programa  que leia  um valor em metros e o exiba convertido em centímetros e milímetros.
 
-metros = int(input('Digite o valor em metros: '))
+metros = float(input('Digite o valor em metros: '))
 centrimetros = metros * 100
 milimetros = metros * 1000
